Ajuste tbla: Paciente, cita, facturas, historia

+++ database/migrations/2020_10_09_173929_create_usuario_rol_table.php
@@ -16,7 +16,7 @@ class CreateUsuarioRolTable extends Migration
         Schema::create('usuario_rol', function (Blueprint $table) {
             $table->unsignedInteger('rol_id');
             $table->foreign('rol_id', 'fk_usuariorol_rol')->references('id')->on('rol')->onDelete('restrict')->onUpdate('restrict');
-            $table->unsignedInteger('usuario_id');
+            $table->unsignedBigInteger('usuario_id');
             $table->foreign('usuario_id', 'fk_usuariorol_usuario')->references('id')->on('usuario')->onDelete('restrict')->onUpdate('restrict');
             $table->timestamps();
         });
